Refactor mesh classes to submodule (`.mesh`)
