+++ nanomesh/_mesh_shared.py
@@ -7,7 +7,7 @@ from typing import Union
 import numpy as np
 
 from .image import Plane, Volume
-from .mesh import BaseMesh
+from .mesh._base import BaseMesh
 
 logger = logging.getLogger(__name__)
