--- nanomesh/mesh.py
@@ -1,590 +1,0 @@
-from __future__ import annotations
-
-from types import MappingProxyType
-from typing import TYPE_CHECKING, Dict, List, Optional, Sequence
-
-import matplotlib.pyplot as plt
-import meshio
-import numpy as np
-import pyvista as pv
-import scipy
-
-from .region_markers import RegionMarker, RegionMarkerLike
-
-if TYPE_CHECKING:
-    import open3d
-
-    from .mesh_container import MeshContainer
-
-
-class PruneZ0Mixin:
-
-    def prune_z_0(self):
-        """Drop third dimension (z) coordinates if present and all values are
-        equal to 0 (within tolerance).
-
-        For compatibility, sometimes a column with zeroes is added, for
-        example when exporting to gmsh2.2 format. This method drops that
-        column.
-        """
-        TOL = 1e-9
-
-        is_3_dimensional = self.points.shape[1] == 3
-        if not is_3_dimensional:
-            return
-
-        z_close_to_0 = np.all(np.abs(self.points[:, 2]) < TOL)
-        if z_close_to_0:
-            self.points = self.points[:, 0:2]
-
-
-class BaseMesh:
-    cell_type: str = 'base'
-
-    def __init__(self,
-                 points: np.ndarray,
-                 cells: np.ndarray,
-                 fields: Dict[str, int] = None,
-                 region_markers: List[RegionMarker] = None,
-                 **cell_data):
-        """Base class for meshes.
-
-        Parameters
-        ----------
-        points : (m, n) np.ndarray[float]
-            Array with points.
-        cells : (i, j) np.ndarray[int]
-            Index array describing the cells of the mesh.
-        fields : Dict[str, int]:
-            Mapping from field names to labels
-        region_markers : List[RegionMarker], optional
-            List of region markers used for assigning labels to regions.
-            Defaults to an empty list.
-        **cell_data
-            Additional cell data. Argument must be a 1D numpy array
-            matching the number of cells defined by `i`.
-        """
-        default_key = 'physical'
-        if (not cell_data) or (default_key in cell_data):
-            self.default_key = default_key
-        else:
-            self.default_key = list(cell_data.keys())[0]
-
-        self.fields = dict(fields) if fields else {}
-
-        self.points = points
-        self.cells = cells
-        self.field_to_number = MappingProxyType(self.fields)
-        self.region_markers = [] if region_markers is None else region_markers
-        self.cell_data = cell_data
-
-    def __repr__(self, indent: int = 0):
-        """Canonical string representation."""
-        region_markers = set(m.name if m.name else m.label
-                             for m in self.region_markers)
-        s = (
-            f'{self.__class__.__name__}(',
-            f'    points = {self.points.shape},',
-            f'    cells = {self.cells.shape},',
-            f'    fields = {tuple(self.field_to_number.keys())},',
-            f'    region_markers = {tuple(region_markers)},',
-            f'    cell_data = {tuple(self.cell_data.keys())},',
-            ')',
-        )
-
-        prefix = ' ' * indent
-        return f'\n{prefix}'.join(s)
-
-    @property
-    def number_to_field(self):
-        """Mapping from numbers to fields, proxy to `.field_to_number`."""
-        return MappingProxyType(
-            {v: k
-             for k, v in self.field_to_number.items()})
-
-    def add_region_marker(self, region_marker: RegionMarkerLike):
-        """Add marker to list of region markers.
-
-        Parameters
-        ----------
-        region_marker : RegionMarkerLike
-            Either a `RegionMarker` object or `(label, point)` tuple,
-            where the label must be an `int` and the point a 2- or
-            3-element numpy array.
-        """
-        if not isinstance(region_marker, RegionMarker):
-            region_marker = RegionMarker(*region_marker)
-
-        self.region_markers.append(region_marker)
-
-    def add_region_markers(self, region_markers: Sequence[RegionMarkerLike]):
-        """Add marker to list of region markers.
-
-        Parameters
-        ----------
-        region_markers : List[RegionMarkerLike]
-            List of region markers passed to `.add_region_marker`.
-        """
-        for region_marker in region_markers:
-            self.add_region_marker(region_marker)
-
-    def to_meshio(self) -> 'meshio.Mesh':
-        """Return instance of `meshio.Mesh`."""
-        cells = [
-            (self.cell_type, self.cells),
-        ]
-
-        mesh = meshio.Mesh(self.points, cells)
-
-        for key, value in self.cell_data.items():
-            mesh.cell_data[key] = [value]
-
-        return mesh
-
-    @classmethod
-    def from_meshio(cls, mesh: 'meshio.Mesh'):
-        """Return `BaseMesh` from meshio object."""
-        points = mesh.points
-        cells = mesh.cells[0].data
-        cell_data = {}
-
-        for key, value in mesh.cell_data.items():
-            # PyVista chokes on ':ref' in cell_data
-            key = key.replace(':ref', '-ref')
-            cell_data[key] = value[0]
-
-        return BaseMesh.create(points=points, cells=cells, **cell_data)
-
-    @classmethod
-    def create(cls, points, cells, **cell_data):
-        """Class dispatcher."""
-        cell_dimensions = cells.shape[1]
-        if cell_dimensions == 2:
-            item_class = LineMesh
-        elif cell_dimensions == 3:
-            item_class = TriangleMesh
-        elif cell_dimensions == 4:
-            item_class = TetraMesh
-        else:
-            item_class = cls
-        return item_class(points=points, cells=cells, **cell_data)
-
-    def write(self, *args, **kwargs):
-        """Simple wrapper around `meshio.write`."""
-        self.to_meshio().write(*args, **kwargs)
-
-    @classmethod
-    def read(cls, filename, **kwargs):
-        """Simple wrapper around `meshio.read`."""
-        mesh = meshio.read(filename, **kwargs)
-        return cls.from_meshio(mesh)
-
-    def to_pyvista_unstructured_grid(self) -> 'pv.PolyData':
-        """Return instance of `pyvista.UnstructuredGrid`.
-
-        References
-        ----------
-        https://docs.pyvista.org/core/point-grids.html#pv-unstructured-grid-class-methods
-        """
-        return pv.from_meshio(self.to_meshio())
-
-    def plot(self, **kwargs):
-        raise NotImplementedError
-
-    def plot_mpl(self, **kwargs):
-        raise NotImplementedError
-
-    def plot_itk(self, **kwargs):
-        """Wrapper for `pyvista.plot_itk`.
-
-        Parameters
-        ----------
-        **kwargs
-            Extra keyword arguments passed to `pyvista.plot_itk`
-        """
-        return pv.plot_itk(self.to_meshio(), **kwargs)
-
-    def plot_pyvista(self, **kwargs):
-        """Wrapper for `pyvista.plot`.
-
-        Parameters
-        ----------
-        **kwargs
-            Extra keyword arguments passed to `pyvista.plot`
-        """
-        return pv.plot(self.to_meshio(), **kwargs)
-
-    @property
-    def cell_centers(self):
-        """Return centers of cells (mean of corner points)."""
-        return self.points[self.cells].mean(axis=1)
-
-    def get_cell_data(self, key: str, default_value: float = 0) -> np.ndarray:
-        """Get cell data with optional default value.
-
-        Parameters
-        ----------
-        key : str
-            Key of the cell data to retrieve.
-        default_value : float, optional
-            Optional default value (if cell data does not exist)
-
-        Returns
-        -------
-        np.ndarray
-        """
-        try:
-            return self.cell_data[self.default_key]
-        except KeyError:
-            return np.ones(len(self.cells), dtype=int) * default_value
-
-    @property
-    def zero_labels(self):
-        """Return zero labels as fallback."""
-        return np.zeros(len(self.cells), dtype=int)
-
-    @property
-    def labels(self):
-        """Shortcut for cell labels."""
-        try:
-            return self.cell_data[self.default_key]
-        except KeyError:
-            return self.zero_labels
-
-    @labels.setter
-    def labels(self, data: np.ndarray):
-        """Shortcut for setting cell labels."""
-        self.cell_data[self.default_key] = data
-
-    @property
-    def dimensions(self):
-        """Return number of dimensions for point data."""
-        return self.points.shape[1]
-
-    def reverse_cell_order(self):
-        """Reverse order of cells and cell data."""
-        self.cells = self.cells[::-1]
-        for key, data in self.cell_data.items():
-            self.cell_data[key] = data[::-1]
-
-
-class LineMesh(BaseMesh):
-    cell_type = 'line'
-
-    def plot(self, *args, **kwargs):
-        """Shortcut for `.plot_mpl`"""
-        return self.plot_mpl(*args, **kwargs)
-
-    def plot_mpl(self, *args, **kwargs) -> plt.Axes:
-        """Simple line mesh plot using `matplotlib`. See
-        `.plotting.linemeshplot` for details.
-
-        Parameters
-        ----------
-        *args
-            Arguments passed to `.plotting.linemeshplot`
-        **kwargs
-            Keyword arguments passed to `.plotting.linemeshplot`
-
-        Returns
-        -------
-        plt.Axes
-        """
-        from .plotting import linemeshplot
-        return linemeshplot(self, *args, **kwargs)
-
-    def label_boundaries(self,
-                         left: Optional[int | str] = None,
-                         right: Optional[int | str] = None,
-                         top: Optional[int | str] = None,
-                         bottom: Optional[int | str] = None,
-                         key: str = None):
-        """Labels the boundaries of the mesh with the given value.
-
-        Parameters
-        ----------
-        left : int | str, optional
-            Labels left boundary segments with the given value. If a string
-            is passed, the `.fields` attribute is updated with the
-            field / value pair.
-        right : int | str, optional
-            Same as above.
-        top : int | str, optional
-            Same as above.
-        bottom : int | str, optional
-            Same as above.
-        key : str, optional
-            Key of the `.cell_data` dictionary to update. Defaults to
-            `.default_key`.
-        """
-        if not key:
-            key = self.default_key
-
-        for side, f_bound, col in (
-            (left, np.min, 1),
-            (right, np.max, 1),
-            (top, np.max, 0),
-            (bottom, np.min, 0),
-        ):
-            if not side:
-                continue
-
-            bound = f_bound(self.points)
-            idx = np.argwhere(self.points[:, col] == bound)
-            side_idx = np.nonzero(np.all(np.isin(self.cells, idx), axis=1))
-
-            if isinstance(side, str):
-                int_label = max(self.cell_data[key].max(), 1) + 1
-                self.fields[side] = int_label
-            else:
-                int_label = side
-
-            self.cell_data[key][side_idx] = int_label
-
-    def triangulate(self, opts: str = 'pq30Aa100') -> MeshContainer:
-        """Triangulate mesh using `triangle`."""
-        from ._triangle_wrapper import triangulate
-        points = self.points
-        segments = self.cells
-        regions = [(m.point[0], m.point[1], m.label, m.constraint)
-                   for m in self.region_markers]
-
-        segment_markers = self.cell_data.get('segment_markers', None)
-
-        mesh = triangulate(points=points,
-                           segments=segments,
-                           regions=regions,
-                           segment_markers=segment_markers,
-                           opts=opts)
-
-        fields = {m.label: m.name for m in self.region_markers if m.name}
-        mesh.set_field_data('triangle', fields)
-        return mesh
-
-
-class TriangleMesh(BaseMesh, PruneZ0Mixin):
-    cell_type = 'triangle'
-
-    def plot(self, **kwargs):
-        """Shortcut for `.plot_mpl` or `.plot_itk` depending on dimensions."""
-        if self.dimensions == 2:
-            return self.plot_mpl(**kwargs)
-        else:
-            return self.plot_itk(**kwargs)
-
-    def plot_mpl(self, *args, **kwargs) -> plt.Axes:
-        """Simple triangle mesh plot using `matplotlib`. See
-        `.plotting.trianglemeshplot` for details.
-
-        Parameters
-        ----------
-        *args
-            Arguments passed to `.plotting.trianglemeshplot`
-        **kwargs
-            Keyword arguments passed to `.plotting.trianglemeshplot`
-
-        Returns
-        -------
-        plt.Axes
-        """
-        from .plotting import trianglemeshplot
-        return trianglemeshplot(self, *args, **kwargs)
-
-    def to_trimesh(self):
-        """Return instance of `trimesh.Trimesh`."""
-        import trimesh
-        return trimesh.Trimesh(vertices=self.points, faces=self.cells)
-
-    def to_open3d(self) -> 'open3d.geometry.TriangleMesh':
-        """Return instance of `open3d.geometry.TriangleMesh`."""
-        import open3d
-        return open3d.geometry.TriangleMesh(
-            vertices=open3d.utility.Vector3dVector(self.points),
-            triangles=open3d.utility.Vector3iVector(self.cells))
-
-    def to_polydata(self) -> 'pv.PolyData':
-        """Return instance of `pyvista.Polydata`."""
-        points = self.points
-        cells = self.cells
-        # preprend 3 to indicate number of points per cell
-        stacked_cells = np.hstack(np.insert(cells, 0, values=3,
-                                            axis=1))  # type: ignore
-        return pv.PolyData(points, stacked_cells, n_faces=len(cells))
-
-    @classmethod
-    def from_open3d(cls, mesh: 'open3d.geometry.TriangleMesh') -> TriangleMesh:
-        """Return instance of `TriangleMesh` from open3d."""
-        points = np.asarray(mesh.vertices)
-        cells = np.asarray(mesh.triangles)
-        return cls(points=points, cells=cells)
-
-    @classmethod
-    def from_scipy(cls, mesh: 'scipy.spatial.qhull.Delaunay') -> TriangleMesh:
-        """Return instance of `TriangleMesh` from `scipy.spatial.Delaunay`
-        object."""
-        points = mesh.points
-        cells = mesh.simplices
-        return cls(points=points, cells=cells)
-
-    @classmethod
-    def from_trimesh(cls, mesh) -> TriangleMesh:
-        """Return instance of `TriangleMesh` from trimesh."""
-        return cls(points=mesh.vertices, cells=mesh.faces)
-
-    @classmethod
-    def from_triangle_dict(cls, dct: dict) -> TriangleMesh:
-        """Return instance of `TriangleMesh` from triangle results dict."""
-        from .mesh_container import MeshContainer
-        mesh = MeshContainer.from_triangle_dict(dct)
-        return mesh.get('triangle')
-
-    def optimize(self,
-                 *,
-                 method='CVT (block-diagonal)',
-                 tol: float = 1.0e-3,
-                 max_num_steps: int = 10,
-                 **kwargs) -> TriangleMesh:
-        """Optimize mesh using `optimesh`.
-
-        Parameters
-        ----------
-        method : str, optional
-            Method name
-        tol : float, optional
-            Tolerance
-        max_num_steps : int, optional
-            Maximum number of optimization steps.
-        **kwargs
-            Arguments to pass to `optimesh.optimize_points_cells`
-
-        Returns
-        -------
-        TriangleMesh
-        """
-        import optimesh
-        points, cells = optimesh.optimize_points_cells(
-            X=self.points,
-            cells=self.cells,
-            method=method,
-            tol=tol,
-            max_num_steps=max_num_steps,
-            **kwargs,
-        )
-        return TriangleMesh(points=points, cells=cells)
-
-    def tetrahedralize(self, **kwargs) -> 'TetraMesh':
-        """Tetrahedralize a contour.
-
-        Parameters
-        ----------
-        **kwargs
-            Keyword arguments passed to `nanomesh.tetrahedralize`.
-
-        Returns
-        -------
-        mesh : TetraMesh
-            Tetrahedralized mesh.
-        """
-        from ._tetgen_wrapper import tetrahedralize
-        mesh = tetrahedralize(self, **kwargs)
-        return mesh
-
-
-class TetraMesh(BaseMesh):
-    cell_type = 'tetra'
-
-    def to_open3d(self):
-        """Return instance of `open3d.geometry.TetraMesh`."""
-        import open3d
-        return open3d.geometry.TetraMesh(
-            vertices=open3d.utility.Vector3dVector(self.points),
-            tetras=open3d.utility.Vector4iVector(self.cells))
-
-    @classmethod
-    def from_open3d(cls, mesh) -> TetraMesh:
-        """Return instance of `TetraMesh` from open3d."""
-        points = np.asarray(mesh.vertices)
-        cells = np.asarray(mesh.tetras)
-        return cls(points=points, cells=cells)
-
-    @classmethod
-    def from_pyvista_unstructured_grid(cls, grid: 'pv.UnstructuredGrid'):
-        """Return infance of `TetraMesh` from `pyvista.UnstructuredGrid`."""
-        assert grid.cells[0] == 4
-        cells = grid.cells.reshape(grid.n_cells, 5)[:, 1:]
-        points = np.array(grid.points)
-        return cls(points=points, cells=cells)
-
-    def plot(self, **kwargs):
-        """Shortcut for `.plot_pyvista`."""
-        return self.plot_pyvista(**kwargs)
-
-    def plot_pyvista(self, **kwargs):
-        """Show grid using `pyvista`.
-
-        Parameters
-        ----------
-        **kwargs
-            Keyword arguments passed to `pyvista.Plotter().add_mesh`.
-        """
-        return self.to_pyvista_unstructured_grid().plot(**kwargs)
-
-    def plot_submesh(
-        self,
-        index: int = None,
-        along: str = 'x',
-        invert: bool = False,
-        show: bool = True,
-        backend: str = None,
-        **kwargs,
-    ):
-        """Show submesh using `pyvista`.
-
-        Parameters
-        ----------
-        index : int, optional
-            Index of where to cut the mesh. Shows all tetrahedra
-            with cell center < index. Picks the half-way
-            point along the axis by default.
-        along : str, optional
-            Direction along which to cut.
-        invert : bool, optional
-            Invert the cutting operation, and show all tetrahedra with
-            cell center > index.
-        show : bool, optional
-            If true, show the plot
-        **kwargs:
-            Keyword arguments passed to `pyvista.Plotter().add_mesh`.
-
-        plotter : `pyvista.Plotter`
-            Return plotter instance.
-        """
-        grid = self.to_pyvista_unstructured_grid()
-
-        # get cell centroids
-        cells = grid.cells.reshape(-1, 5)[:, 1:]
-        cell_center = grid.points[cells].mean(1)
-
-        # extract cells below index
-        axis = 'zyx'.index(along)
-
-        if index is None:
-            # pick half-way point
-            i, j = axis * 2, axis * 2 + 2
-            index = np.mean(grid.bounds[i:j])
-
-        mask = cell_center[:, axis] < index
-
-        if invert:
-            mask = ~mask
-
-        cell_ind = mask.nonzero()[0]
-        subgrid = grid.extract_cells(cell_ind)
-
-        plotter = pv.Plotter()
-        plotter.add_mesh(subgrid, **kwargs)
-
-        if show:
-            plotter.show(jupyter_backend=backend)
-
-        return plotter
